fix for issue 667
