+++ src/reporter/tests/utils.py
@@ -2,6 +2,7 @@ from conftest import QL_URL
 from datetime import datetime, timedelta, timezone
 import json
 import requests
+from requests import Response
 import time
 from typing import Callable, Iterable, List, Optional, Tuple, Union
 import re
@@ -106,6 +107,15 @@ def send_notifications(service, notifications, service_path='/'):
     for n in notifications:
         r = requests.post(notify_url(), data=json.dumps(n), headers=h)
         assert r.ok
+    return r
+
+
+def insert_entities(entities: Union[List[dict], dict],
+                    service: str = None, service_path: str = None,
+                    expected_status_code=200) -> Response:
+    response = send_notifications(service, entities)
+    assert response.status_code == expected_status_code
+    return response
 
 
 def send_notifications_different_types(service, notifications):
